xml/parsing/manakai/extperef-2.dat

#data
<!DOCTYPE a [
  <!ENTITY % foo "<?xml encoding='utf-8' ?><!ENTITY bar 'aa'>">
  %foo;
]>
<a b="&bar;x"/>
#errors
1;1;no XML decl;;;s
2;19;xml:dtd:pi;;;w
2;44;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a>
|   <?xml encoding='utf-8' ?>
| <{} a>
|   b="aax"
